Pull urlValidation logic to core from junitSupport
- This will be used by HttpStub to validate baseUrls
